UI controller

+++ bignetcoin-wallettemplate/src/main/java/wallettemplate/CoinModel.java
@@ -1,17 +1,16 @@
 package wallettemplate;
 
-import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleLongProperty;
 
 public class CoinModel {
     private SimpleLongProperty value;
-    private SimpleIntegerProperty tokenid;
+    private SimpleLongProperty tokenid;
 
     public SimpleLongProperty value() {
         return value;
     }
 
-    public SimpleIntegerProperty tokenid() {
+    public SimpleLongProperty tokenid() {
         return tokenid;
     }
 
@@ -23,7 +22,7 @@ public class CoinModel {
         this.value.set(value);
     }
 
-    public int getTokenid() {
+    public long getTokenid() {
         return tokenid.get();
     }
 
@@ -31,10 +30,10 @@ public class CoinModel {
         this.tokenid.set(tokenid);
     }
 
-    public CoinModel(long value, int tokenid) {
+    public CoinModel(long value, long tokenid) {
         super();
         this.value = new SimpleLongProperty(value);
-        this.tokenid = new SimpleIntegerProperty(tokenid);
+        this.tokenid = new SimpleLongProperty(tokenid);
     }
 
 }
